+++ pyobjc-framework-Automator/PyObjCTest/test_automator.py
@@ -17,4 +17,12 @@ class TestAutomator(TestCase):
 
 class TestCallableMetadata(TestCase):
     def test_callable_metadata_is_sane(self):
-        self.assertCallableMetadataIsSane(Automator)
+        self.assertCallableMetadataIsSane(
+            Automator,
+            exclude_attrs={
+                (
+                    "NSObject",
+                    "copyRenderedTextureForCGLContext_pixelFormat_bounds_isFlipped_",
+                )
+            },
+        )