+++ pyobjc-framework-ContactsUI/PyObjCTest/test_contactsui.py
@@ -5,4 +5,12 @@ import ContactsUI
 
 class TestCallableMetadata(TestCase):
     def test_callable_metadata_is_sane(self):
-        self.assertCallableMetadataIsSane(ContactsUI)
+        self.assertCallableMetadataIsSane(
+            ContactsUI,
+            exclude_attrs={
+                (
+                    "NSObject",
+                    "copyRenderedTextureForCGLContext_pixelFormat_bounds_isFlipped_",
+                ),
+            },
+        )