+++ pyobjc-framework-GameCenter/PyObjCTest/test_gamecenter.py
@@ -5,4 +5,12 @@ import GameCenter
 
 class TestCallableMetadata(TestCase):
     def test_callable_metadata_is_sane(self):
-        self.assertCallableMetadataIsSane(GameCenter)
+        self.assertCallableMetadataIsSane(
+            GameCenter,
+            exclude_attrs={
+                (
+                    "NSObject",
+                    "copyRenderedTextureForCGLContext_pixelFormat_bounds_isFlipped_",
+                ),
+            },
+        )