+++ pyobjc-framework-GameKit/PyObjCTest/test_gamekit.py
@@ -5,4 +5,12 @@ import GameKit
 
 class TestCallableMetadata(TestCase):
     def test_callable_metadata_is_sane(self):
-        self.assertCallableMetadataIsSane(GameKit)
+        self.assertCallableMetadataIsSane(
+            GameKit,
+            exclude_attrs={
+                (
+                    "NSObject",
+                    "copyRenderedTextureForCGLContext_pixelFormat_bounds_isFlipped_",
+                ),
+            },
+        )