+++ pyobjc-framework-IMServicePlugIn/PyObjCTest/test_imserviceplugin.py
@@ -5,4 +5,12 @@ import IMServicePlugIn
 
 class TestCallableMetadata(TestCase):
     def test_callable_metadata_is_sane(self):
-        self.assertCallableMetadataIsSane(IMServicePlugIn)
+        self.assertCallableMetadataIsSane(
+            IMServicePlugIn,
+            exclude_attrs={
+                (
+                    "NSObject",
+                    "copyRenderedTextureForCGLContext_pixelFormat_bounds_isFlipped_",
+                ),
+            },
+        )