+++ pyobjc-framework-Quartz/PyObjCTest/test_quartz.py
@@ -5,4 +5,10 @@ import Quartz
 
 class TestCallableMetadata(TestCase):
     def test_callable_metadata_is_sane(self):
-        self.assertCallableMetadataIsSane(Quartz)
+        self.assertCallableMetadataIsSane(
+            Quartz,
+            exclude_attrs={
+                "NSColor",
+                ("NSColor", "scn_C3DColorIgnoringColorSpace_success_"),
+            },
+        )