+++ pyobjc-framework-ScreenSaver/PyObjCTest/test_screensaver.py
@@ -49,4 +49,12 @@ class TestScreenSaver(TestCase):
 
 class TestCallableMetadata(TestCase):
     def test_callable_metadata_is_sane(self):
-        self.assertCallableMetadataIsSane(ScreenSaver)
+        self.assertCallableMetadataIsSane(
+            ScreenSaver,
+            exclude_attrs={
+                (
+                    "NSObject",
+                    "copyRenderedTextureForCGLContext_pixelFormat_bounds_isFlipped_",
+                )
+            },
+        )